requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-chroma==0.1.4
langserve==0.3.0
langgraph==0.2.20
pydantic>=2.7.4
python-dotenv==1.0.1
pypdf==4.3.1
wikipedia==1.4.0
faiss-cpu==1.7.4
streamlit==1.34.0
tavily-python==0.5.1
openai>=1.40.0
sentence-transformers==2.0.0

